requirements.txt
absl-py==1.4.0
certifi==2022.12.7
charset-normalizer==3.0.1
cloudpickle==2.2.1
contourpy==1.0.7
cvxpy==1.3.0
cycler==0.11.0
decorator==4.4.2
dill==0.3.6
ecos==2.0.12
fonttools==4.38.0
glfw==2.5.9
gymnasium==0.27.1
gymnasium-notices==0.0.1
gymnasium-robotics==1.2.0
idna==3.4
imageio==2.25.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.0.0
importlib-resources==5.12.0
jax-jumpy==0.2.0
keyboard==0.13.5
kiwisolver==1.4.4
matplotlib==3.7.0
moviepy==1.0.3
mujoco==2.3.2
multiprocess==0.70.14
numpy==1.23.5
opencv-python==4.8.0.76
osqp==0.6.2.post8
packaging==23.0
pandas==1.5.3
pathos==0.3.0
Pillow==9.4.0
pox==0.3.2
ppft==1.7.6.6
proglog==0.1.10
PyOpenGL==3.1.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
qdldl==0.1.5.post3
requests==2.28.2
scipy==1.10.1
scs==3.2.2
seaborn==0.12.2
six==1.16.0
tqdm==4.64.1
typing-extensions==4.5.0
urllib3==1.26.14
zipp==3.15.0
